5.19. Больше подробностей (часть 2). Правки после код-ревью
