+++ src/utils/func.ts
@@ -1,3 +1,4 @@
+import {maxNearOffers} from '../const';
 import {TCity, TLocationCoordinates, TOffer, TOffersByCity} from '../types';
 
 export function getRandomInt(min: number, max: number): number {
@@ -35,4 +36,20 @@ export function getOffersByCity(offers: TOffer[]): TOffersByCity[] {
   });
 
   return offersByCity;
+}
+
+export function getNearOffers(offers: TOffer[], curOffer: TOffer): TOffer[] {
+  const nearOffers = [];
+  for (const offer of offers) {
+    if (
+      offer.city === curOffer.city
+      && offer.id !== curOffer.id
+      && nearOffers.length < maxNearOffers
+    ) {
+      nearOffers.push(offer);
+    } else if (nearOffers.length === maxNearOffers) {
+      break;
+    }
+  }
+  return nearOffers;
 }
